+++ configmanager.h
@@ -314,11 +314,11 @@ class ConfigManager
 
 		const std::string& getString(uint32_t _what) const;
 		bool getBool(uint32_t _what) const;
-		int32_t getNumber(uint32_t _what) const;
+		int64_t getNumber(uint32_t _what) const;
 		double getDouble(uint32_t _what) const;
 
 		bool setString(uint32_t _what, const std::string& _value);
-		bool setNumber(uint32_t _what, int32_t _value);
+		bool setNumber(uint32_t _what, int64_t _value);
 		bool setBool(uint32_t _what, bool _value);
 
 		void getValue(const std::string& key, lua_State* _L) {LuaInterface::getValue(key, L, _L);}
@@ -334,7 +334,7 @@ class ConfigManager
 		{
 			return LuaInterface::getGlobalBool(L, _identifier, _default);
 		}
-		int32_t getGlobalNumber(const std::string& _identifier, const int32_t _default = 0)
+		int64_t getGlobalNumber(const std::string& _identifier, const int64_t _default = 0)
 		{
 			return LuaInterface::getGlobalNumber(L, _identifier, _default);
 		}
@@ -348,7 +348,7 @@ class ConfigManager
 
 		std::string m_confString[LAST_STRING_CONFIG];
 		bool m_confBool[LAST_BOOL_CONFIG];
-		int32_t m_confNumber[LAST_NUMBER_CONFIG];
+		int64_t m_confNumber[LAST_NUMBER_CONFIG];
 		double m_confDouble[LAST_DOUBLE_CONFIG];
 };
 #endif
